Remove old static constructor `FreshBooksApi::init()` as it not longer serves a purpose
When the change was made to add a proper constructor that takes the domain and token as parameters, it made a breaking change in the code, that the method was no longer passed in to the constructor. The update also made the `init()` method and protected static properties `$_domain` and `$_token` obsolete as well.